fix: Remove GCM from cinterop — CommonCrypto GCM APIs are SPI
CommonCrypto's GCM mode (kCCModeGCM, CCCryptorGCMOneshotEncrypt)
is System Programming Interface, not public API. Removed from
cinterop def file.

SHA-256, MD5, and random bytes work via public CC_SHA256/CC_MD5/
CCRandomGenerateBytes APIs.

AES encrypt/decrypt left as TODO — requires CryptoKit Swift bridge
(planned for future wave). AES tests moved to JVM-only (jvmTest).
