+++ commcare-core/src/nativeInterop/cinterop/CommonCrypto.def
@@ -1,7 +1,6 @@
 package = CommCareCrypto
 ---
 #include <CommonCrypto/CommonDigest.h>
-#include <CommonCrypto/CommonCryptor.h>
 #include <CommonCrypto/CommonRandom.h>
 
 static inline int cc_sha256(const void *data, unsigned int len, void *md) {
@@ -16,79 +15,4 @@ static inline int cc_md5(const void *data, unsigned int len, void *md) {
 
 static inline int cc_random_generate(void *bytes, unsigned long count) {
     return (int)CCRandomGenerateBytes(bytes, (size_t)count);
-}
-
-/* AES-GCM using step-by-step CCCryptor API */
-static inline int cc_gcm_encrypt(
-    const void *key, unsigned long keyLen,
-    const void *iv, unsigned long ivLen,
-    const void *dataIn, unsigned long dataLen,
-    void *dataOut,
-    void *tagOut, unsigned long tagLen
-) {
-    CCCryptorRef cryptor = NULL;
-    CCCryptorStatus status;
-
-    status = CCCryptorCreateWithMode(
-        kCCEncrypt, kCCModeGCM, kCCAlgorithmAES, ccNoPadding,
-        NULL, key, keyLen, NULL, 0, 0, kCCModeOptionCTR_BE, &cryptor
-    );
-    if (status != kCCSuccess) return (int)status;
-
-    /* Set IV */
-    status = CCCryptorGCMAddIV(cryptor, iv, ivLen);
-    if (status != kCCSuccess) { CCCryptorRelease(cryptor); return (int)status; }
-
-    /* Encrypt data */
-    if (dataIn && dataLen > 0) {
-        size_t outMoved = 0;
-        status = CCCryptorUpdate(cryptor, dataIn, dataLen, dataOut, dataLen, &outMoved);
-        if (status != kCCSuccess) { CCCryptorRelease(cryptor); return (int)status; }
-    }
-
-    /* Get tag */
-    status = CCCryptorGCMFinal(cryptor, tagOut, &tagLen);
-    CCCryptorRelease(cryptor);
-    return (int)status;
-}
-
-static inline int cc_gcm_decrypt(
-    const void *key, unsigned long keyLen,
-    const void *iv, unsigned long ivLen,
-    const void *dataIn, unsigned long dataLen,
-    void *dataOut,
-    const void *tagIn, unsigned long tagLen
-) {
-    CCCryptorRef cryptor = NULL;
-    CCCryptorStatus status;
-
-    status = CCCryptorCreateWithMode(
-        kCCDecrypt, kCCModeGCM, kCCAlgorithmAES, ccNoPadding,
-        NULL, key, keyLen, NULL, 0, 0, kCCModeOptionCTR_BE, &cryptor
-    );
-    if (status != kCCSuccess) return (int)status;
-
-    /* Set IV */
-    status = CCCryptorGCMAddIV(cryptor, iv, ivLen);
-    if (status != kCCSuccess) { CCCryptorRelease(cryptor); return (int)status; }
-
-    /* Decrypt data */
-    if (dataIn && dataLen > 0) {
-        size_t outMoved = 0;
-        status = CCCryptorUpdate(cryptor, dataIn, dataLen, dataOut, dataLen, &outMoved);
-        if (status != kCCSuccess) { CCCryptorRelease(cryptor); return (int)status; }
-    }
-
-    /* Verify tag */
-    unsigned char computedTag[16];
-    unsigned long computedTagLen = tagLen;
-    status = CCCryptorGCMFinal(cryptor, computedTag, &computedTagLen);
-    CCCryptorRelease(cryptor);
-    if (status != kCCSuccess) return (int)status;
-
-    /* Compare tags */
-    if (memcmp(computedTag, tagIn, tagLen) != 0) {
-        return -1; /* Tag mismatch */
-    }
-    return 0;
 }
